@@ -51,7 +51,7 @@ class KeyFileManager {
 }
 
 class KeygenPage extends StatefulWidget {
-  KeygenPage({Key key, this.title}) : super(key: key);
+  KeygenPage({Key? key, required this.title}) : super(key: key);
   final String title;
 
   @override
