@@ -3,7 +3,7 @@ import 'package:loc_chain_app/pages/home.dart';
 import 'package:loc_chain_app/pages/keygen.dart';
 
 class NavBarWidget extends StatefulWidget {
-  NavBarWidget({Key key}) : super(key: key);
+  NavBarWidget({Key? key}) : super(key: key);
 
   @override
   _NavBarWidgetState createState() => _NavBarWidgetState();
@@ -11,7 +11,7 @@ class NavBarWidget extends StatefulWidget {
 
 class _NavBarWidgetState extends State<NavBarWidget> {
   int _selectedIndex = 0;
-  StatefulWidget _selectedWidget;
+  Widget _selectedWidget = HomePage(title: 'Home');
   static List<Widget> pages = <Widget>[
     HomePage(
       title: 'Home',
